+++ include/pyquantlib/pyquantlib.h
@@ -292,6 +292,10 @@ namespace ql_instruments {
     void fixedratebond(py::module_&);
     void zerocouponbond(py::module_&);
     void floatingratebond(py::module_&);
+    void amortizingfixedratebond(py::module_&);
+    void amortizingfloatingratebond(py::module_&);
+    void cmsratebond(py::module_&);
+    void cpibond(py::module_&);
     void swap(py::module_&);
     void fixedvsfloatingswap(py::module_&);
     void vanillaswap(py::module_&);